Fix for underlying from/to problem
The fromCards() / toCards() methods in HolsDerGeierUtil were using this.from instead of Main.getMain().from, leading to the class values being used to reset the players. In the first fix they were already set to their correct defaults, but were still not reacting to overrides in the options menu since the methods, and therefore also the player reset, referenced not the instanced values but the non-overwritten generic class defaults.